isDocumentRetrieval: move useEffect to TypingBox
